no more utils remove deprecated tests
