+++ btk/measure.py
@@ -37,6 +37,77 @@ class Measurement_params(ABC):
         return None
 
 
+class Basic_measure_params(Measurement_params):
+    """Class to perform detection by identifying peaks with skimage"""
+
+    @staticmethod
+    def get_centers(image):
+        """Return centers detected when object detection is performed on the
+        input image with skimage.feature.peak_local_max.
+
+        Args:
+            image (np.ndarray): Image (single band) of galaxy to perform measurement
+
+        Returns:
+                centers: x and y coordinates of detected  centroids
+        """
+        # set detection threshold to 5 times std of image
+        threshold = 5 * np.std(image)
+        coordinates = skimage.feature.peak_local_max(
+            image, min_distance=2, threshold_abs=threshold
+        )
+        return np.stack((coordinates[:, 1], coordinates[:, 0]), axis=1)
+
+    def get_deblended_images(self, data, index):
+        """Returns scarlet modeled blend and centers for the given blend"""
+        image = np.mean(data["blend_images"][index], axis=2)
+        peaks = self.get_centers(image)
+        return {"deblend_image": None, "peaks": peaks}
+
+
+class SEP_params(Measurement_params):
+    """Class to perform detection and deblending with SEP"""
+
+    def __init__(self):
+        self.catalog = None
+        self.segmentation = None
+
+    def get_centers(self, image):
+        """Return centers detected when object detection and photometry
+        is done on input image with SEP.
+        It also initializes the self.catalog and self.segmentation attributes
+        of the class object.
+        Args:
+            image: Image (single band) of galaxy to perform measurement on.
+        Returns:
+            centers: x and y coordinates of detected  centroids
+        """
+        sep = __import__("sep")
+        bkg = sep.Background(image)
+        self.catalog, self.segmentation = sep.extract(
+            image, 1.5, err=bkg.globalrms, segmentation_map=True
+        )
+        centers = np.stack((self.catalog["x"], self.catalog["y"]), axis=1)
+        return centers
+
+    def get_deblended_images(self, data, index):
+        """Performs SEP detection on the band-coadd image and returns the
+        detected peaks.
+        Args:
+            data (dict): Output generated by btk.draw_blends containing blended
+                images, isolated images, observing conditions and blend
+                catalog, for a given batch.
+            index (int): Index number of blend scene in the batch to preform
+                measurement on.
+        Returns:
+            dict with the centers of sources detected by SEP detection
+            algorithm.
+        """
+        image = np.mean(data["blend_images"][index], axis=2)
+        peaks = self.get_centers(image)
+        return {"deblend_image": None, "peaks": peaks}
+
+
 class MeasureGenerator:
     def __init__(
         self,
